requirements.txt
python-dotenv
assemblyai
flask